pyproject.toml
[project]
name = "databricks-ai-steward"
version = "0.1.0"
description = "MCP server that lets AI agents interact with Databricks safely."
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.2.0",
    "python-dotenv>=1.0.0",
    "databricks-sdk>=0.30",
    "sqlglot>=25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    # Required by stress/probe_langchain_compat.py — verifies the
    # langchain-ai/langchain-mcp-adapters wrapper can load our tools.
    "langchain-mcp-adapters>=0.1",
]
http = [
    # Required for --transport sse and --transport streamable-http.
    # FastMCP pulls these in transitively today, but listing them
    # explicitly makes `pip install '.[http]'` the documented contract.
    "uvicorn>=0.27",
    "starlette>=0.36",
]

[tool.hatch.build.targets.wheel]
packages = ["mcp_server"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
